feat(remark): Split out plugins
Our Remark plugins now have individual modules and are found under the
`remark` directory. There are next to no functional changes other than
the plugins receiving more descriptive names and the IMG and SVG plugins
being merged into one.

This is groundwork for the next plugin which will merge adjacent code
blocks.
