+++ remark/spreadListItems.test.ts
@@ -1,0 +1,39 @@
+import { spreadListItems } from './spreadListItems';
+
+describe('spreadListItems', () => {
+  it('spreads all list items', () => {
+    const ast = {
+      type: 'list',
+      children: [
+        {
+          type: 'listItem',
+          spread: true,
+        },
+        {
+          type: 'listItem',
+          spread: false,
+        },
+      ],
+    } as const;
+
+    const astCopy = { ...ast };
+
+    spreadListItems()(astCopy);
+
+    expect(astCopy).toMatchInlineSnapshot(`
+      Object {
+        "children": Array [
+          Object {
+            "spread": true,
+            "type": "listItem",
+          },
+          Object {
+            "spread": true,
+            "type": "listItem",
+          },
+        ],
+        "type": "list",
+      }
+    `);
+  });
+});
